# web/packages/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Premise.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
  "paths": {
...
    "/auth/login": {
      "get": {
        "tags": [
          "Premise.Api"
        ],
        "parameters": [
          {
            "name": "returnUrl",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hint",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "org",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/callback": {
      "get": {
        "tags": [
          "Premise.Api"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/auth/switch-org": {
      "post": {
        "tags": [
          "Premise.Api"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchOrgRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SwitchOrgRequest": {
        "required": [
          "orgId"
        ],
        "type": "object",
        "properties": {
          "orgId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    }
  },
...
}